Refactor hooks imports to improve module organization and clarity
- Updated imports in various components to source hooks from their respective directories under "@/components".
- Introduced a new hook for fetching project conversation counts in the hooks index file.
- Removed unused hooks from the query module to streamline functionality.
