+++ echo/frontend/src/lib/query.ts
@@ -11,31 +11,22 @@ import {
 import { useI18nNavigate } from "@/hooks/useI18nNavigate";
 import {
   api,
-  checkUnsubscribeStatus,
   getLatestProjectAnalysisRunByProjectId,
   getProjectChatContext,
-  getProjectConversationCounts,
-  getResourceById,
 } from "./api";
 import { toast } from "@/components/common/Toaster";
 import { directus } from "./directus";
 import {
   Query,
   aggregate,
-  passwordRequest,
-  passwordReset,
   readItem,
   readItems,
   readUser,
-  registerUser,
-  registerUserVerify,
   updateItem,
-  deleteItem,
 } from "@directus/sdk";
-import { ADMIN_BASE_URL } from "@/config";
 
 // always throws a error with a message
-function throwWithMessage(e: unknown): never {
+export function throwWithMessage(e: unknown): never {
   if (
     e &&
     typeof e === "object" &&
@@ -137,132 +128,6 @@ export const useCreateProjectMutation = () => {
   });
 };
 
-// todo: add redirection logic here
-export const useLoginMutation = () => {
-  return useMutation({
-    mutationFn: (payload: Parameters<typeof directus.login>) => {
-      return directus.login(...payload);
-    },
-    onSuccess: () => {
-      toast.success("Login successful");
-    },
-  });
-};
-
-export const useRegisterMutation = () => {
-  const navigate = useI18nNavigate();
-  return useMutation({
-    mutationFn: async (payload: Parameters<typeof registerUser>) => {
-      try {
-        const response = await directus.request(registerUser(...payload));
-        return response;
-      } catch (e) {
-        try {
-          throwWithMessage(e);
-        } catch (inner) {
-          if (inner instanceof Error) {
-            if (inner.message === "You don't have permission to access this.") {
-              throw new Error(
-                "Oops! It seems your email is not eligible for registration at this time. Please consider joining our waitlist for future updates!",
-              );
-            }
-          }
-        }
-      }
-    },
-    onSuccess: () => {
-      toast.success("Please check your email to verify your account.");
-      navigate("/check-your-email");
-    },
-    onError: (e) => {
-      toast.error(e.message);
-    },
-  });
-};
-
-export const useVerifyMutation = (doRedirect: boolean = true) => {
-  const navigate = useI18nNavigate();
-
-  return useMutation({
-    mutationFn: async (data: { token: string }) => {
-      try {
-        const response = await directus.request(registerUserVerify(data.token));
-        return response;
-      } catch (e) {
-        throwWithMessage(e);
-      }
-    },
-    onSuccess: () => {
-      toast.success("Email verified successfully.");
-      if (doRedirect) {
-        setTimeout(() => {
-          // window.location.href = `/login?new=true`;
-          navigate(`/login?new=true`);
-        }, 4500);
-      }
-    },
-    onError: (e) => {
-      toast.error(e.message);
-    },
-  });
-};
-
-export const useRequestPasswordResetMutation = () => {
-  const navigate = useI18nNavigate();
-  return useMutation({
-    mutationFn: async (email: string) => {
-      try {
-        const response = await directus.request(
-          passwordRequest(email, `${ADMIN_BASE_URL}/password-reset`),
-        );
-        return response;
-      } catch (e) {
-        throwWithMessage(e);
-      }
-    },
-    onSuccess: () => {
-      toast.success("Password reset email sent successfully");
-      navigate("/check-your-email");
-    },
-    onError: (e) => {
-      toast.error(e.message);
-    },
-  });
-};
-
-export const useResetPasswordMutation = () => {
-  const navigate = useI18nNavigate();
-  return useMutation({
-    mutationFn: async ({
-      token,
-      password,
-    }: {
-      token: string;
-      password: string;
-    }) => {
-      try {
-        const response = await directus.request(passwordReset(token, password));
-        return response;
-      } catch (e) {
-        throwWithMessage(e);
-      }
-    },
-    onSuccess: () => {
-      toast.success(
-        "Password reset successfully. Please login with new password.",
-      );
-      navigate("/login");
-    },
-    onError: (e) => {
-      try {
-        toast.error(e.message);
-      } catch (e) {
-        toast.error("Error resetting password. Please contact support.");
-      }
-    },
-  });
-};
-
 export const useLogoutMutation = () => {
   const queryClient = useQueryClient();
   const navigate = useI18nNavigate();
@@ -335,13 +200,6 @@ export const useUpdateProjectByIdMutation = () => {
   });
 };
 
-export const useResourceById = (resourceId: string) => {
-  return useQuery({
-    queryKey: ["resources", resourceId],
-    queryFn: () => getResourceById(resourceId),
-  });
-};
-
 export const useConversationById = ({
   conversationId,
   loadConversationChunks = false,
@@ -376,27 +234,6 @@ export const useConversationById = ({
         }),
       ),
     ...useQueryOpts,
-  });
-};
-
-export const useDeleteConversationChunkByIdMutation = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: (chunkId: string) =>
-      directus.request(deleteItem("conversation_chunk", chunkId)),
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: ["conversations"],
-      });
-    },
-  });
-};
-
-export const useProjectConversationCounts = (projectId: string) => {
-  return useQuery({
-    queryKey: ["projects", projectId, "conversation-counts"],
-    queryFn: () => getProjectConversationCounts(projectId),
-    refetchInterval: 15000,
   });
 };
 
@@ -611,20 +448,5 @@ export const useProjectReportViews = (reportId: number) => {
       };
     },
     refetchInterval: 30000,
-  });
-};
-
-export const useCheckUnsubscribeStatus = (token: string, projectId: string) => {
-  return useQuery<{ eligible: boolean }>({
-    queryKey: ["checkUnsubscribe", token, projectId],
-    queryFn: async () => {
-      if (!token || !projectId) {
-        throw new Error("Invalid or missing unsubscribe link.");
-      }
-      const response = await checkUnsubscribeStatus(token, projectId);
-      return response.data;
-    },
-    retry: false,
-    refetchOnWindowFocus: false,
   });
 };
