+++ echo/frontend/src/routes/auth/RequestPasswordReset.tsx
@@ -1,6 +1,6 @@
 import { t } from "@lingui/core/macro";
 import { Trans } from "@lingui/react/macro";
-import { useRequestPasswordResetMutation } from "@/lib/query";
+import { useRequestPasswordResetMutation } from "@/components/auth/hooks";
 import { Button, Container, Stack, TextInput, Title } from "@mantine/core";
 import { useDocumentTitle } from "@mantine/hooks";
 import { useForm } from "react-hook-form";
